+++ src/bc_parse.c
@@ -79,6 +79,7 @@ static bool bc_parse_inst_isLeaf(BcInst t) {
  * that can legally end a statement. In bc's case, it could be a newline, a
  * semicolon, and a brace in certain cases.
  * @param p  The parser.
+ * @return   True if the token is a legal delimiter.
  */
 static bool bc_parse_isDelimiter(const BcParse *p) {
 
@@ -125,6 +126,23 @@ static bool bc_parse_isDelimiter(const BcParse *p) {
 	}
 
 	return good;
+}
+
+/**
+ * Returns true if we are in top level of a function body. The POSIX grammar
+ * is defined such that anything is allowed after a function body, so we must
+ * use this function to detect that case when ending a function body.
+ * @param p  The parser.
+ * @return   True if we are in the top level of parsing a function body.
+ */
+static bool bc_parse_TopFunc(const BcParse *p) {
+
+	bool good = p->flags.len == 2;
+
+	uint16_t val = BC_PARSE_FLAG_BRACE | BC_PARSE_FLAG_FUNC_INNER;
+	val |= BC_PARSE_FLAG_FUNC;
+
+	return good && BC_PARSE_TOP_FLAG(p) == val;
 }
 
 /**
@@ -881,7 +899,7 @@ static void bc_parse_endBody(BcParse *p, bool brace) {
 		bc_lex_next(&p->l);
 
 		// If the next token is not a delimiter, that is a problem.
-		if (BC_ERR(!bc_parse_isDelimiter(p)))
+		if (BC_ERR(!bc_parse_isDelimiter(p) && !bc_parse_TopFunc(p)))
 			bc_parse_err(p, BC_ERR_PARSE_TOKEN);
 	}
 
@@ -1741,6 +1759,14 @@ static void bc_parse_stmt(BcParse *p) {
 
 	// Make sure semicolons are eaten.
 	while (p->l.t == BC_LEX_SCOLON) bc_lex_next(&p->l);
+
+	// POSIX's grammar does not allow a function definition after a semicolon
+	// without a newline, so check specifically for that case and error if
+	// the POSIX standard flag is set.
+	if (p->l.last == BC_LEX_SCOLON && p->l.t == BC_LEX_KW_DEFINE && BC_IS_POSIX)
+	{
+		bc_parse_err(p, BC_ERR_POSIX_FUNC_AFTER_SEMICOLON);
+	}
 }
 
 void bc_parse_parse(BcParse *p) {